move to styles folder and rename the base function
