--- src/libs/getModalStyles/index.android.js
@@ -1,6 +1,0 @@
-import getModalStyles from './getModalStyles';
-
-export default (type, windowDimensions, popoverAnchorPosition = {}, containerStyle = {}) => ({
-    ...getModalStyles(type, windowDimensions, popoverAnchorPosition, containerStyle),
-    shouldAddTopSafeAreaPadding: false,
-});
