--- src/libs/getModalStyles/index.js
@@ -1,3 +1,0 @@
-import getModalStyles from './getModalStyles';
-
-export default getModalStyles;
