+++ src/styles/getModalStyles/index.android.js
@@ -1,0 +1,6 @@
+import baseGetModalStyles from './baseGetModalStyles';
+
+export default (type, windowDimensions, popoverAnchorPosition = {}, containerStyle = {}) => ({
+    ...baseGetModalStyles(type, windowDimensions, popoverAnchorPosition, containerStyle),
+    shouldAddTopSafeAreaPadding: false,
+});
