+++ src/styles/getModalStyles/index.js
@@ -1,0 +1,3 @@
+import baseGetModalStyles from './baseGetModalStyles';
+
+export default baseGetModalStyles;
